Updated Main and Process Classes
The Process class can now create deep copies of process arrays and reset
them back to their original state. This is so everyone receives the same
list of processes. The Main class just has some test cases in it right
now to make sure the copying and resetting are working correctly.